conver all the presigned upload models to the new fileupload response body
