+++ src/lib/models/index.ts
@@ -11,6 +11,7 @@ export * from './group-session-models';
 export * from './institution';
 export * from './integrated-care-models';
 export * from './interaction-models';
+export * from './presigned-upload-models';
 export * from './provider';
 export * from './reference-data';
 export * from './roles';
