+++ src/lib/models/presigned-upload-models.tsx
@@ -1,0 +1,18 @@
+import { Method } from 'axios';
+
+export interface PresignedUploadModel {
+	accessUrl: string;
+	contentType: string;
+	expirationTimestamp: string;
+	expirationTimestampDescription: string;
+	httpHeaders: Record<string, string>;
+	httpMethod: Method;
+	url: string;
+}
+
+export interface PresignedUploadResponse {
+	fileUploadResult: {
+		fileUploadId: string;
+		presignedUpload: PresignedUploadModel;
+	};
+}
